Separate Lens ListView & Rename Lens ViewModel to LensList ViewModel
