+++ lens_review_ios/ContentView.swift
@@ -12,11 +12,15 @@ struct ContentView: View
 {
     
 //    @ObservedObject var lensViewModel: LensViewModel = LensViewModel()
-    @EnvironmentObject var lensViewModel: LensViewModel
+    @EnvironmentObject var lensViewModel: LensListViewModel
 
     var body: some View {
         TabView{
             LensListView()
+                .tabItem {
+                    Image(systemName: "list.dash")
+                    Text("렌즈 리스트")
+                }
             Text("2")
                 .tabItem {
                     Image(systemName: "list.dash")
@@ -28,32 +32,6 @@ struct ContentView: View
 
 struct ContentView_Previews: PreviewProvider {
     static var previews: some View {
-        ContentView().environmentObject(LensViewModel())
-    }
-}
-
-struct LensListView: View {
-    
-    @EnvironmentObject var lensViewModel: LensViewModel
-    
-    var body: some View {
-        NavigationView
-        {
-            List
-            {
-                ForEach(lensViewModel.lensList) { lens_ in
-                    HStack {
-                        NavigationLink(destination: LensDetailView(selectedLensId: lens_.id))
-                        {
-                            Text(lens_.name)
-                        }
-                    }
-                }
-            }
-            .navigationBarTitle(Text("Lens List"))
-        }.tabItem {
-            Image(systemName: "list.dash")
-            Text("렌즈 리스트")
-        }
+        ContentView().environmentObject(LensListViewModel())
     }
 }
