+++ lens_review_ios/LensListViewModel.swift
@@ -8,7 +8,7 @@
 
 import Combine
 
-class LensViewModel: ObservableObject
+class LensListViewModel: ObservableObject
 {
     @Published var lensList = [Lens]()
